refactor: Move DomainUrl et al to ontol-core, drops ontol-examples dependency on ontol-compiler
