+++ ontol/ontol-core/Cargo.toml
@@ -13,10 +13,12 @@ doctest = false
 [dependencies]
 ontol-macros.path = "../ontol-macros"
 arcstr.workspace = true
+async-trait = "0.1"
 bit-vec = { version = "0.8", features = ["serde"] }
 compact_str = { version = "0.8", features = ["serde"] }
 serde = { version = "1", features = ["derive"] }
 thin-vec.workspace = true
+url = "2"
 
 [dev-dependencies]
 pretty_assertions = "1"
