+++ ontol/ontol-core/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod debug;
 pub mod property;
 pub mod rustdoc;
+pub mod url;
 
 extern crate self as ontol_core;
