+++ ontol/ontol-examples/Cargo.toml
@@ -13,6 +13,6 @@ rust-version.workspace = true
 doctest = false
 
 [dependencies]
-ontol-compiler.path = "../ontol-compiler"
+ontol-core.path = "../ontol-core"
 async-trait.workspace = true
 url = "2"
